Added include of <sys/socket.h> in globals.h.
This fixes a compilation error in Linux, since MSG_NOSIGNAL is defined in
<sys/socket.h> and was thus invisible from globals.h, conflicting then with the
definition given by the inclusion of <sys/socket.h> in xdcc.h.

I considered moving the ifdef to xdcc.c but would imply further copy-paste of
the conditional definition if I wanted to use MSG_NOSIGNAL in other files, such
as network.c or irc.c.